fix: add lrt & move input box
